requirements.txt
streamlit
pdfplumber
pytesseract
Pillow
scikit-learn
google-generativeai
groq
indic-nlp-library
sentencepiece
torch
scipy
transformers
datasets
accelerate
numpy
pandas